refactor: densify Slice C orchestrator and phase-planner AI tests
Topic-split remaining >300-line domain-orchestrator and phase-planner pin suites so AC4 can proceed without changing planner behavior.
