+++ packages/colonizethis_ai/test/planning/domain_planner_orchestrator_cross_planner_build_cases.dart
@@ -1,0 +1,146 @@
+// Build-selection cases for domain_planner_orchestrator_cross_planner_test.dart (Refs #4602).
+
+import 'package:colonizethis_test/test.dart';
+import 'package:colonizethis_ai/colonizethis_ai.dart';
+import 'package:colonizethis_data/colonizethis_data.dart';
+import 'package:colonizethis_logic/colonizethis_logic.dart';
+import 'package:colonizethis_models/colonizethis_models.dart';
+
+import '../support/domain_planner_test_fake_api.dart';
+import '../support/planner_test_helpers.dart';
+
+void registerDomainPlannerOrchestratorCrossPlannerBuildCases() {
+  test('with strongCargo and ship candidate picks ship deterministically', () {
+    const regimentBuild = BuildUnitOrder(
+      unitType: 'peasant_levies',
+      isMilitary: true,
+      spawnProvinceId: 'oldWorld|p1',
+    );
+    const shipBuild = BuildUnitOrder(
+      unitType: 'fluyte',
+      isMilitary: false,
+      spawnProvinceId: 'oldWorld|p1',
+    );
+    const fakeApi = FakeOrderSuggestionAPIForDomainPlannerTests(
+      work: [],
+      build: [regimentBuild, shipBuild],
+      move: [],
+      research: [],
+      navalMove: [],
+      navalMission: [],
+      diplomatic: [],
+    );
+    const economyPlanStrongCargo = EconomyPlan(
+      productionAssignments: [],
+      cargoPreference: CargoPreference.strongCargo,
+    );
+    final game = Game(
+      id: 'g1',
+      worldState: WorldState(
+        turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 1),
+        oldWorld: const RegionData(),
+        newWorld: const RegionData(),
+      ),
+      players: const [
+        Player(id: 'gp1', displayName: 'A', isHuman: false, leaderKey: 'henry'),
+      ],
+    );
+    const topology = MapTopology(nodes: [], edges: []);
+    final view = buildPlayerView(game, topology, 'gp1');
+    const postQuotaSnapshot = AIWorldSnapshot(
+      playerId: 'gp1',
+      threats: ThreatSummary(),
+      opportunities: OpportunitySummary(),
+      conquest: ConquestSummary(
+        oldWorldProvincesOwned: kObserverConquestMinOwProvincesPerGp + 2,
+      ),
+      economy: EconomySummary(),
+      relations: {},
+    );
+    final orders = runDomainPlannersInTest(
+      game: game,
+      topology: topology,
+      view: view,
+      snapshot: postQuotaSnapshot,
+      turnSeed: 42,
+      primaryGoal: StrategicGoal.expand,
+      config: const AIConfig(
+        leaderId: 'henry',
+        personalityId: 'henry',
+        hiddenAgendaId: 'peacemaker',
+      ),
+      suggestionAPI: fakeApi,
+      economyPlan: economyPlanStrongCargo,
+    );
+
+    final builds = orders.buildUnitOrdersByPlayerId['gp1'] ?? [];
+    expect(builds.length, 1);
+    expect(
+      builds.single.unitType,
+      'fluyte',
+      reason: 'strongCargo should favour cargo ship over regiment',
+    );
+  });
+
+  test(
+    'build selection is deterministic for same seed and cargoPreference none',
+    () {
+      const regimentBuild = BuildUnitOrder(
+        unitType: 'peasant_levies',
+        isMilitary: true,
+        spawnProvinceId: 'oldWorld|p1',
+      );
+      const shipBuild = BuildUnitOrder(
+        unitType: 'fluyte',
+        isMilitary: false,
+        spawnProvinceId: 'oldWorld|p1',
+      );
+      const fakeApi = FakeOrderSuggestionAPIForDomainPlannerTests(
+        work: [],
+        build: [regimentBuild, shipBuild],
+        move: [],
+        research: [],
+        navalMove: [],
+        navalMission: [],
+        diplomatic: [],
+      );
+      final game = Game(
+        id: 'g1',
+        worldState: WorldState(
+          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 1),
+          oldWorld: const RegionData(),
+          newWorld: const RegionData(),
+        ),
+        players: const [
+          Player(
+            id: 'gp1',
+            displayName: 'A',
+            isHuman: false,
+            leaderKey: 'victoria',
+          ),
+        ],
+      );
+      const topology = MapTopology(nodes: [], edges: []);
+      final orders1 = runDomainPlannersInTest(
+        game: game,
+        topology: topology,
+        turnSeed: 999,
+        suggestionAPI: fakeApi,
+      );
+      final orders2 = runDomainPlannersInTest(
+        game: game,
+        topology: topology,
+        turnSeed: 999,
+        suggestionAPI: fakeApi,
+      );
+
+      final build1 = orders1.buildUnitOrdersByPlayerId['gp1']?.single.unitType;
+      final build2 = orders2.buildUnitOrdersByPlayerId['gp1']?.single.unitType;
+      expect(
+        build1,
+        build2,
+        reason: 'same seed and economyPlan should yield same build choice',
+      );
+    },
+  );
+}
