+++ packages/colonizethis_ai/test/planning/phase_planner_declare_war_targets_test.dart
@@ -1,45 +1,12 @@
-// Unit tests for `phase_planner_declare_war_targets.dart`
-// (Refs #2509 S5 orchestrator adapter slice).
-//
-// Adapter contract pinned here (from
-// `SPEC/ai/phase-planner-dispatch.md` § Adapter helpers):
-//
-//   gpExpandDeclareWarTargetFromPhasePlan(outcome):
-//     - EXPAND       -> outcome.expandDeclareWarTargetFactionId
-//     - COLONIAL-lite -> outcome.expandDeclareWarTargetFactionId
-//       (OW push keeps running during the safeguard)
-//     - COLONIAL     -> null
-//     - DEVELOP      -> null
-//
-//   gpColonialDeclareWarTargetFromPhasePlan(outcome):
-//     - COLONIAL with acquisition method == declareWar
-//                    -> outcome.colonialAcquisitionTarget.targetFactionId
-//     - COLONIAL with acquisition method == joinEmpire / purchaseLand
-//                    -> null
-//     - COLONIAL with null acquisition target
-//                    -> null
-//     - EXPAND / COLONIAL-lite / DEVELOP
-//                    -> null (structural; colonialAcquisitionTarget is
-//                            suppressed outside COLONIAL by the dispatcher)
-//
-// Fixtures here construct `PhasePlanOutcome` instances directly so the
-// tests do not require a `Game` / `AIWorldSnapshot` pair. Outcome
-// composition from real `runPhasePlanners` dispatches is already covered
-// by `phase_planner_dispatch_test.dart`.
+// Adapter helpers: `SPEC/ai/phase-planner-dispatch.md` (Refs #2509 S5, #4602).
 
-import 'package:colonizethis_ai/src/planning/colonial_phase_planner.dart';
 import 'package:colonizethis_ai/src/planning/observer_goal_phase.dart';
 import 'package:colonizethis_ai/src/planning/phase_planner_declare_war_targets.dart';
 import 'package:colonizethis_ai/src/planning/phase_planner_dispatch.dart';
-import 'package:colonizethis_ai/src/planning/phase_priority_weights.dart';
 import 'package:colonizethis_test/test.dart';
 
-const PhasePriorityWeights _nwAcquisitionZeroExpand = PhasePriorityWeights(
-  oldWorldConquest: 0.95,
-  newWorldAcquisition: 0.0,
-  oldWorldCivilian: 0.90,
-  newWorldCivilian: 0.10,
-);
+import 'phase_planner_declare_war_targets_colonial_cases.dart';
+import 'phase_planner_declare_war_targets_exclusion_cases.dart';
 
 void main() {
   group('gpExpandDeclareWarTargetFromPhasePlan', () {
@@ -120,216 +87,6 @@ void main() {
     });
   });
 
-  group('gpColonialDeclareWarTargetFromPhasePlan', () {
-    test('COLONIAL + declareWar method surfaces targetFactionId', () {
-      const outcome = PhasePlanOutcome(
-        phase: ObserverGoalPhase.colonial,
-        colonialAcquisitionTarget: ColonialAcquisitionTarget(
-          targetFactionId: 'tribe1',
-          method: AcquisitionMethod.declareWar,
-        ),
-      );
-      expect(gpColonialDeclareWarTargetFromPhasePlan(outcome), 'tribe1');
-    });
-
-    test('COLONIAL + joinEmpire method surfaces null', () {
-      const outcome = PhasePlanOutcome(
-        phase: ObserverGoalPhase.colonial,
-        colonialAcquisitionTarget: ColonialAcquisitionTarget(
-          targetFactionId: 'tribe2',
-          method: AcquisitionMethod.joinEmpire,
-        ),
-      );
-      expect(
-        gpColonialDeclareWarTargetFromPhasePlan(outcome),
-        isNull,
-        reason:
-            'Join Empire acquisitions resolve via establishOverture, not '
-            'declareWar; the adapter must surface null so the diplomacy '
-            'planner does not declare war when the acquisition path is '
-            'peaceful.',
-      );
-    });
-
-    test('COLONIAL + purchaseLand method surfaces null', () {
-      const outcome = PhasePlanOutcome(
-        phase: ObserverGoalPhase.colonial,
-        colonialAcquisitionTarget: ColonialAcquisitionTarget(
-          targetFactionId: 'tribe3',
-          method: AcquisitionMethod.purchaseLand,
-        ),
-      );
-      expect(
-        gpColonialDeclareWarTargetFromPhasePlan(outcome),
-        isNull,
-        reason:
-            'purchase_land acquisitions resolve via the build pipeline, '
-            'not declareWar.',
-      );
-    });
-
-    test('COLONIAL + null acquisition target surfaces null', () {
-      const outcome = PhasePlanOutcome(phase: ObserverGoalPhase.colonial);
-      expect(
-        gpColonialDeclareWarTargetFromPhasePlan(outcome),
-        isNull,
-        reason:
-            'Null acquisition means no method fired this turn; the '
-            'military / naval pair fall back to the at-war arm without a '
-            'fresh declareWar.',
-      );
-    });
-
-    test('EXPAND with newWorldAcquisition=0 surfaces null even when '
-        'acquisition target is non-null', () {
-      const outcome = PhasePlanOutcome(
-        phase: ObserverGoalPhase.expand,
-        colonialAcquisitionTarget: ColonialAcquisitionTarget(
-          targetFactionId: 'tribe4',
-          method: AcquisitionMethod.declareWar,
-        ),
-        priorityWeights: _nwAcquisitionZeroExpand,
-      );
-      expect(gpColonialDeclareWarTargetFromPhasePlan(outcome), isNull);
-    });
-
-    test('EXPAND with newWorldAcquisition>0 surfaces declareWar target '
-        '(Refs #2847)', () {
-      const outcome = PhasePlanOutcome(
-        phase: ObserverGoalPhase.expand,
-        colonialAcquisitionTarget: ColonialAcquisitionTarget(
-          targetFactionId: 'tribe4',
-          method: AcquisitionMethod.declareWar,
-        ),
-      );
-      expect(
-        gpColonialDeclareWarTargetFromPhasePlan(outcome),
-        'tribe4',
-      );
-    });
-
-    test(
-      'COLONIAL-lite surfaces null even when acquisition target is non-null',
-      () {
-        const outcome = PhasePlanOutcome(
-          phase: ObserverGoalPhase.colonialLite,
-          colonialAcquisitionTarget: ColonialAcquisitionTarget(
-            targetFactionId: 'tribe5',
-            method: AcquisitionMethod.declareWar,
-          ),
-        );
-        expect(
-          gpColonialDeclareWarTargetFromPhasePlan(outcome),
-          isNull,
-          reason:
-              'COLONIAL-lite is the EXPAND safeguard that suppresses NW '
-              'declareWar / joinEmpire / purchase_land. Even if a future '
-              'regression populated colonialAcquisitionTarget under the '
-              'safeguard, the adapter must keep returning null.',
-        );
-      },
-    );
-
-    test('DEVELOP surfaces null even when acquisition target is non-null', () {
-      const outcome = PhasePlanOutcome(
-        phase: ObserverGoalPhase.develop,
-        colonialAcquisitionTarget: ColonialAcquisitionTarget(
-          targetFactionId: 'tribe6',
-          method: AcquisitionMethod.declareWar,
-        ),
-      );
-      expect(gpColonialDeclareWarTargetFromPhasePlan(outcome), isNull);
-    });
-
-    test('defaultColonial outcome surfaces null', () {
-      expect(
-        gpColonialDeclareWarTargetFromPhasePlan(
-          PhasePlanOutcome.defaultColonial,
-        ),
-        isNull,
-      );
-    });
-
-    test('defaultExpand outcome surfaces null', () {
-      expect(
-        gpColonialDeclareWarTargetFromPhasePlan(PhasePlanOutcome.defaultExpand),
-        isNull,
-      );
-    });
-  });
-
-  group('adapter mutual exclusion', () {
-    test('EXPAND surfaces only the EXPAND adapter target', () {
-      const outcome = PhasePlanOutcome(
-        phase: ObserverGoalPhase.expand,
-        expandDeclareWarTargetFactionId: 'minorE',
-      );
-      expect(gpExpandDeclareWarTargetFromPhasePlan(outcome), 'minorE');
-      expect(gpColonialDeclareWarTargetFromPhasePlan(outcome), isNull);
-    });
-
-    test('COLONIAL surfaces only the COLONIAL adapter target', () {
-      const outcome = PhasePlanOutcome(
-        phase: ObserverGoalPhase.colonial,
-        colonialAcquisitionTarget: ColonialAcquisitionTarget(
-          targetFactionId: 'tribeC',
-          method: AcquisitionMethod.declareWar,
-        ),
-      );
-      expect(gpExpandDeclareWarTargetFromPhasePlan(outcome), isNull);
-      expect(gpColonialDeclareWarTargetFromPhasePlan(outcome), 'tribeC');
-    });
-
-    test('COLONIAL-lite surfaces only the EXPAND adapter target', () {
-      // The COLONIAL-lite safeguard runs EXPAND alongside its overture /
-      // naval directives; no NW declareWar surfaces because acquisition
-      // is structurally suppressed.
-      const outcome = PhasePlanOutcome(
-        phase: ObserverGoalPhase.colonialLite,
-        expandDeclareWarTargetFactionId: 'minorL',
-      );
-      expect(gpExpandDeclareWarTargetFromPhasePlan(outcome), 'minorL');
-      expect(gpColonialDeclareWarTargetFromPhasePlan(outcome), isNull);
-    });
-
-    test('DEVELOP surfaces neither adapter target', () {
-      const outcome = PhasePlanOutcome(
-        phase: ObserverGoalPhase.develop,
-        expandDeclareWarTargetFactionId: 'minorD',
-        colonialAcquisitionTarget: ColonialAcquisitionTarget(
-          targetFactionId: 'tribeD',
-          method: AcquisitionMethod.declareWar,
-        ),
-      );
-      expect(gpExpandDeclareWarTargetFromPhasePlan(outcome), isNull);
-      expect(gpColonialDeclareWarTargetFromPhasePlan(outcome), isNull);
-    });
-  });
-
-  group('determinism (Must-have #7)', () {
-    test('gpExpandDeclareWarTargetFromPhasePlan stable across calls', () {
-      const outcome = PhasePlanOutcome(
-        phase: ObserverGoalPhase.expand,
-        expandDeclareWarTargetFactionId: 'minorStable',
-      );
-      expect(
-        gpExpandDeclareWarTargetFromPhasePlan(outcome),
-        gpExpandDeclareWarTargetFromPhasePlan(outcome),
-      );
-    });
-
-    test('gpColonialDeclareWarTargetFromPhasePlan stable across calls', () {
-      const outcome = PhasePlanOutcome(
-        phase: ObserverGoalPhase.colonial,
-        colonialAcquisitionTarget: ColonialAcquisitionTarget(
-          targetFactionId: 'tribeStable',
-          method: AcquisitionMethod.declareWar,
-        ),
-      );
-      expect(
-        gpColonialDeclareWarTargetFromPhasePlan(outcome),
-        gpColonialDeclareWarTargetFromPhasePlan(outcome),
-      );
-    });
-  });
+  registerPhasePlannerDeclareWarTargetsColonialCases();
+  registerPhasePlannerDeclareWarTargetsExclusionCases();
 }
